File structure & Prediction (#4)
* adds file structure
* fixes prediction code
