+++ tuepa/data/__init__.py
@@ -1,0 +1,1 @@
+from .preprocessing import read_passages
